Add methods to add and remove selected values
`addSelectedValues(values)` takes a value or an array of values and adds them to the selection (if not already selected).

`removeSelectedValues(values)` takes a value or an array of values and removes them from the selection.

If the value is not found then both methods do nothing. Also note: both methods use `querySelector()` to find the relevant values. If one value occurs more than once in the wc then only the first value will be added/removed.