fix: system preference setting mode attribute
The wc did not know if mode was set explicitly by user or was set implicitly by system preference settings. This means that once the wc was implicitly set it could no longer adapt to changes in the system preference.

Now changes through system preference are no longer propegated through the mode-attribute but through a class set on the box container within the wc.

The wc can therefore know if the mode was set explicitly or not. If set explicitly it should not react to changes in system preference but if it was not set explicitly it will.